Put createdAt component into its own folder
